requirements.txt

email-validator>=2.2.0
flask>=3.1.1
gunicorn>=23.0.0
requests>=2.32.3
werkzeug>=3.1.3
flask
gunicorn
requests
werkzeug
email-validator
flask-sqlalchemy
psycopg2-binary
sqlalchemy
reportlab>=4.0.0
qrcode[pil]>=7.4.2
pillow>=10.0.0
